Rename provision_task to janus_cli, split operations out to subcommands.
