+++ janus_server/Cargo.toml
@@ -15,6 +15,7 @@ test-util = ["janus_core/test-util", "dep:lazy_static", "dep:testcontainers"]
 
 [dependencies]
 anyhow = "1"
+assert_matches = "1"
 atty = "0.2"
 base64 = "0.13.0"
 bytes = "1.1.0"
@@ -62,7 +63,6 @@ uuid = { version = "1.1.2", features = ["v4"] }
 warp = { version = "^0.3", features = ["tls"] }
 
 [dev-dependencies]
-assert_matches = "1"
 hex = { version = "0.4.3", features = ["serde"] }
 hyper = "0.14.19"
 janus_server = { path = ".", features = ["test-util"] }
