+++ janus_server/src/bin/janus_cli.rs
@@ -1,0 +1,215 @@
+use anyhow::{Context, Result};
+use deadpool_postgres::Pool;
+use janus_core::time::{Clock, RealClock};
+use janus_server::{
+    binary_utils::{janus_main, BinaryContext, BinaryOptions, CommonBinaryOptions},
+    config::{BinaryConfig, CommonConfig},
+    datastore::{self, Datastore},
+    task::Task,
+};
+use serde::{Deserialize, Serialize};
+use std::{
+    path::{Path, PathBuf},
+    sync::Arc,
+};
+use structopt::StructOpt;
+use tokio::fs;
+use tracing::info;
+
+static SCHEMA: &str = include_str!("../../../db/schema.sql");
+
+#[tokio::main]
+async fn main() -> Result<()> {
+    janus_main::<_, Options, Config, _, _>(RealClock::default(), |ctx| async move {
+        ctx.options.cmd.execute(&ctx).await
+    })
+    .await
+}
+
+async fn write_schema(pool: &Pool) -> Result<()> {
+    info!("Writing database schema");
+    let db_client = pool.get().await.context("couldn't get database client")?;
+    db_client
+        .batch_execute(SCHEMA)
+        .await
+        .context("couldn't write database schema")?;
+    Ok(())
+}
+
+async fn provision_tasks<C: Clock>(datastore: &Datastore<C>, tasks_file: &Path) -> Result<()> {
+    // Read tasks file.
+    info!("Reading tasks file");
+    let tasks: Vec<Task> = {
+        let task_file_contents = fs::read_to_string(tasks_file)
+            .await
+            .with_context(|| format!("couldn't read tasks file {:?}", tasks_file))?;
+        serde_yaml::from_str(&task_file_contents)
+            .with_context(|| format!("couldn't parse tasks file {:?}", tasks_file))?
+    };
+
+    // Write all tasks requested.
+    let tasks = Arc::new(tasks);
+    info!(task_count = tasks.len(), "Writing tasks");
+    datastore
+        .run_tx(|tx| {
+            let tasks = Arc::clone(&tasks);
+            Box::pin(async move {
+                for task in tasks.iter() {
+                    // We attempt to delete the task, but ignore "task not found" errors since
+                    // the task not existing is an OK outcome too.
+                    match tx.delete_task(task.id).await {
+                        Ok(_) | Err(datastore::Error::MutationTargetNotFound) => (),
+                        err => err?,
+                    }
+
+                    tx.put_task(task).await?;
+                }
+                Ok(())
+            })
+        })
+        .await
+        .context("couldn't write tasks")
+}
+
+#[derive(Debug, StructOpt)]
+#[structopt(
+    name = "janus-provision-task",
+    about = "Janus `provision task` command",
+    rename_all = "kebab-case",
+    version = env!("CARGO_PKG_VERSION"),
+)]
+struct Options {
+    #[structopt(flatten)]
+    common: CommonBinaryOptions,
+
+    #[structopt(subcommand)]
+    cmd: Command,
+}
+
+impl BinaryOptions for Options {
+    fn common_options(&self) -> &CommonBinaryOptions {
+        &self.common
+    }
+}
+
+#[derive(Debug, StructOpt)]
+enum Command {
+    WriteSchema,
+    ProvisionTasks {
+        /// A YAML file containing a list of tasks to be written. Existing tasks (matching by task
+        /// ID) will be overwritten.
+        tasks_file: PathBuf,
+    },
+}
+
+impl Command {
+    async fn execute<C: Clock>(&self, ctx: &BinaryContext<C, Options, Config>) -> Result<()> {
+        match self {
+            Command::WriteSchema => write_schema(&ctx.pool).await,
+            Command::ProvisionTasks { tasks_file } => {
+                provision_tasks(&ctx.datastore, tasks_file).await
+            }
+        }
+    }
+}
+
+#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
+struct Config {
+    #[serde(flatten)]
+    common_config: CommonConfig,
+}
+
+impl BinaryConfig for Config {
+    fn common_config(&mut self) -> &mut CommonConfig {
+        &mut self.common_config
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::Config;
+    use janus_core::{
+        message::{Role, TaskId},
+        task::VdafInstance,
+        time::RealClock,
+    };
+    use janus_server::{
+        config::test_util::{
+            generate_db_config, generate_metrics_config, generate_trace_config, roundtrip_encoding,
+        },
+        config::CommonConfig,
+        datastore::test_util::{ephemeral_datastore, ephemeral_db_handle},
+        task::test_util::new_dummy_task,
+    };
+    use std::{collections::HashMap, io::Write};
+    use tempfile::NamedTempFile;
+
+    #[tokio::test]
+    async fn write_schema() {
+        let db_handle = ephemeral_db_handle();
+        let ds = db_handle.datastore(RealClock::default());
+
+        // Verify that the query we will run later returns an error if there is no database schema written.
+        ds.run_tx(|tx| Box::pin(async move { tx.get_tasks().await }))
+            .await
+            .unwrap_err();
+
+        // Run the program logic.
+        super::write_schema(&db_handle.pool()).await.unwrap();
+
+        // Verify that the schema was written (by running a query that would fail if it weren't).
+        ds.run_tx(|tx| Box::pin(async move { tx.get_tasks().await }))
+            .await
+            .unwrap();
+    }
+
+    #[tokio::test]
+    async fn provision_tasks() {
+        let tasks = Vec::from([
+            new_dummy_task(
+                TaskId::random(),
+                VdafInstance::Prio3Aes128Count.into(),
+                Role::Leader,
+            ),
+            new_dummy_task(
+                TaskId::random(),
+                VdafInstance::Prio3Aes128Sum { bits: 64 }.into(),
+                Role::Helper,
+            ),
+        ]);
+
+        let (ds, _db_handle) = ephemeral_datastore(RealClock::default()).await;
+
+        // Write tasks to a temporary file.
+        let mut tasks_file = NamedTempFile::new().unwrap();
+        tasks_file
+            .write_all(serde_yaml::to_string(&tasks).unwrap().as_ref())
+            .unwrap();
+        let tasks_path = tasks_file.into_temp_path();
+
+        // Run the program logic.
+        super::provision_tasks(&ds, &tasks_path).await.unwrap();
+
+        // Verify that the expected tasks were written.
+        let want_tasks: HashMap<_, _> = tasks.into_iter().map(|task| (task.id, task)).collect();
+        let got_tasks = ds
+            .run_tx(|tx| Box::pin(async move { tx.get_tasks().await }))
+            .await
+            .unwrap()
+            .into_iter()
+            .map(|task| (task.id, task))
+            .collect();
+        assert_eq!(want_tasks, got_tasks);
+    }
+
+    #[test]
+    fn roundtrip_config() {
+        roundtrip_encoding(Config {
+            common_config: CommonConfig {
+                database: generate_db_config(),
+                logging_config: generate_trace_config(),
+                metrics_config: generate_metrics_config(),
+            },
+        })
+    }
+}
